- added makePanel util
