+++ src/scenario3/index.js
@@ -1,5 +1,6 @@
 import data_sources from '../js/data_sources';
 import getAssetURLs from '../utils/getAssetURLs';
+import makePanel from '../utils/makePanel';
 
 function ready(fn) {
   // replaces $(document).ready() in jQuery
@@ -73,20 +74,6 @@ ready(async function() {
   }
   generateContentFanData();
 
-
-
-  // This is a duplicated helper that should be consolidated!
-  let makePanel = function(data) {
-    // Build content panels with "data"
-    var panel = document.createElement('a-entity');
-    panel.setAttribute("content-group", "");
-
-    let content = panel.components['content-group'];
-    content.initializeFromData(data);
-
-    return panel
-  }
-
   let $contentFan = document.getElementById("contentFan");
   let contentFan = $contentFan.components.contentfan
   let $tabMenu = document.getElementById('tab-menu');
@@ -120,6 +107,7 @@ ready(async function() {
   })
 
   mainMarkerRef.addEventListener("tilt-forward", (e)=>{ // your code here}
+    // add positive/negative review to data_sources personalized json object
     $reviewMenu.components["tab-menu"].confirmSelectedIndex()
   })
